Adapt to JSON v3
We take the advantage of timestamp for simplicity.

The main difference between JSON v2 and v3 is that v3 has an ISO 8601
date format instead of timestamp in seconds since epoc and that null
value properties are omitted from the JSON-objects. Periods on Products
and PaymentPlans are serialized as a ISO 8601 duration (e.g. "PT1M").

Ref:
https://developer.vimond.com/docs/a-note-on-xml-and-json-formats